add 2 headers and add files to vscode debugger
