+++ list/s21_iterator.cc
@@ -1,0 +1,2 @@
+#include "s21_iterator.h"
+
